Lab 6 completed
Some minor errors were corrected, and some animations were put in place. There seems to be a script level error somewhere with the jumping animations but will have to wait until the future to be corrected. Simple door was added and is tied to the frog chest for now.

WRITE UP included this time