Lab 7 done
the spikes now check to see if what they are colliding with is the player before lowering health. Simple health bar UI will spin up or down with the player health. A large frog runs from the chest when opened prompting the player to chase to get a new key to open the door. All using simple ID behaviors for now as Im not sure how the other prefab works and didnt have the time to tinker. Jump has once again broken as it seems to every time the project is pushed to git. writeup included.