Lab 10 complete
Cutting this one a little close there dude. Anyway, particles systems for this lab are done and implemented. Coins make a burst when collected, the door sprays sparks, the spikes make the player drip red goo that for legal reasons is not blood.

The crazy hot red tentacle nightmare is included, although was not implemented into the lab, as well as the frog storm.

Screenshots and the lab writeup are included with the git push as always.