lab 11 beginning
very jank patrolling goose has been added. Using multiple planes and a basic navmesh and properties. I so sorely wish I could do something more advanced than this BS, as my animated frog NPC functions better. I may want to change the routing just a tad, and maybe make it a flying character with a simple little looping sprite I could draw. Like a ghost goose or something.

This enemy also needs to damage the player and find a way with these limiting prefabs to get it to turn around when it reaches the end of its flight path.

I am also realizing I may need to remove and replace the eddie ewi music for the portfolio version of this game, but hopefully not, we will see.