lab 11 is done
The goose is still here, the bat has been added and somehow actually has the jumping mechanic I wanted to sort out with these basic tools, so that's pretty cool.
(RIP spike crawly sprite)

The bat uses hitboxes to know where it is, and which attack state it is in. It is also parented to an external object to get jumping to work because unity is dumb and wont let you keyframe JUST the Y value

the goose thing just wanders and you get hurt if you touch it, simple as.

Writeup is included as usual. :3