lab 12 is done!
Cinemachine cameras are implemented and functioning as intended. And the level has started with expansion because I can control the camera now!

Right now there is the main room as seen before, a tight hallway that zooms the camera tight to the player for the cramped space, and the beginnings of a new platform section with a mid distance follow camera to allow the player to explore a larger room.

There is also a new prefab to help controlling cameras in the future to be easier.

Writeup is included with the push as always.