lab 13 beginnings
Finally fixed the jump forever and jump animation bugs that have been bothering me the whole semester. Also added a cap to the health bar image update script, so the player can no longer heal above 100% HP. The scene also reloads and resets when the player falls bellow 0 HP, so allthough basic the healthbar does now serve a function.

The healing melon and both of the enemies are updated to use the new image updater, although the spikes need re duping so we don't have to change all manually. So for now only spike prime is updated.